refactor: mv memory related methods into separate file
